fix(planning) : removed compiler warnings and errors

+++ src/planning/src/utils/cone2.cpp
@@ -1,8 +1,7 @@
 #include "../../include/utils/cone2.hpp"
 
-int find_cone(std::vector<Cone> &cones, double x, double y){
-    for (int i = 0; i < cones.size(); i++)
-        if (cones[i].position.x == x && cones[i].position.y == y)
-            return i;
-    return -1;
+int find_cone(std::vector<Cone> &cones, double x, double y) {
+  for (int i = 0; i < static_cast<int>(cones.size()); i++)
+    if (cones[i].position.x == x && cones[i].position.y == y) return i;
+  return -1;
 }
